refactor(edit-collective): BankTransfers edit
