+++ lang/es.json
@@ -1118,7 +1118,6 @@
   "paymentmethod.label": "Método de Pago",
   "paymentMethod.manual.edit.description": "Contributors will be able to choose \"Bank Transfer\" as a payment method when they check out. The instructions to make the wire transfer will be emailed to them along with a unique order id. Once you received the money, you will be able to mark the corresponding pending order as paid in your host dashboard.",
   "paymentMethod.manual.edit.description.pricing": "There is no platform fee for donations made this way. However, we ask you to kindly subscribe to our special plans for fiscal hosts to be able to maintain and improve this feature over time (the first $1,000 of yearly budget are included in the free plan)",
-  "paymentMethod.manual.edit.title": "Enable contributors to make donations by wire transfer",
   "paymentMethod.monthlyLimitPerMember.description": "Set a monthly limit to allow other team members of your Organization to use this credit card within that limit. If set to zero, only Organization admins will be able to use this card.",
   "paymentMethod.monthlyLimitPerMember.label": "Limite mensual por miembro",
   "paymentMethod.success": "¡Añadido con éxito!",
@@ -1139,7 +1138,7 @@
   "paymentMethods.manual.edit": "Editar los datos de tu cuenta bancaria",
   "paymentMethods.manual.HowDoesItWork": "¿Cómo funciona?",
   "paymentMethods.manual.instructions": "Instrucciones para transferencia bancaria",
-  "paymentMethods.manual.instructions.default": "Por favor, realiza una transferencia bancaria como se indica a continuación:\n\n<code>\n    Amount: {amount}\n    Reference/Communication: {OrderId}\n    IBAN/Account Number: BE61310126985517\n    BIC: TRIOBEBB\n    Bank name: Triodos\n</code>\n        \nTomará algunos días procesar tu pago.",
+  "paymentMethods.manual.instructions.default": "Please make a bank transfer as follows:\n\n<code>\n    Amount: {amount}\n    Reference/Communication: {OrderId}\n    IBAN/Account Number: BE61310126985517\n    BIC: TRIOBEBB\n    Bank name: Triodos\n</code>\n\nPlease note that it will take a few days to process your payment.",
   "paymentMethods.manual.instructions.title": "Define las instrucciones para realizar una transferencia bancaria a tu cuenta",
   "paymentMethods.manual.upgradePlan": "Subscribe to our special plans for hosts",
   "paymentMethods.removeConfirm": "¿Realmente desea eliminar este método de pago de su cuenta?",
