+++ lang/fr.json
@@ -1118,7 +1118,6 @@
   "paymentmethod.label": "Moyen de paiement",
   "paymentMethod.manual.edit.description": "Les contributeurs seront en mesure de choisir \"Virement Bancaire\" comme moyen de paiement au moment du paiement. Les instructions pour effectuer le virement bancaire leur seront envoyées par courriel avec un numéro de commande unique. Une fois que vous avez reçu l'argent, vous serez en mesure de marquer la commande correspondante comme payée dans votre tableau de bord.",
   "paymentMethod.manual.edit.description.pricing": "Il n'y a pas de frais de plateforme pour les dons effectués de cette façon. Cependant, nous vous demandons de bien vouloir souscrire à nos plans spéciaux pour les hôtes fiscaux afin de pouvoir maintenir et améliorer cette fonctionnalité au fil du temps (les premiers 1,000$ de budget annuel sont inclus dans le plan gratuit)",
-  "paymentMethod.manual.edit.title": "Autorise les contributeurs à effectuer des donations par transfert bancaire",
   "paymentMethod.monthlyLimitPerMember.description": "Fixez une limite mensuelle pour permettre aux autres membres de votre Organisation d'utiliser cette carte de crédit dans cette limite. Si la valeur est zéro, seuls les administrateurs de l'Organisation pourront utiliser cette carte.",
   "paymentMethod.monthlyLimitPerMember.label": "Limite mensuelle par membre",
   "paymentMethod.success": "Ajouté avec succès!",
@@ -1139,7 +1138,7 @@
   "paymentMethods.manual.edit": "Éditer les détails de votre compte en banque",
   "paymentMethods.manual.HowDoesItWork": "Comment ça marche ?",
   "paymentMethods.manual.instructions": "Instructions pour le transfert bancaire",
-  "paymentMethods.manual.instructions.default": "Veuillez effectuer un virement bancaire comme suit :\n\n<code>\n    Montant: {amount}\n    Référence/Communication: {OrderId}\n    IBAN/Numéro de compte: BE61310126985517\n    BIC: TRIOBEBB\n    Nom de la banque: Triodos\n</code>\n\nVeuillez noter que le traitement de votre paiement prendra quelques jours.",
+  "paymentMethods.manual.instructions.default": "Please make a bank transfer as follows:\n\n<code>\n    Amount: {amount}\n    Reference/Communication: {OrderId}\n    IBAN/Account Number: BE61310126985517\n    BIC: TRIOBEBB\n    Bank name: Triodos\n</code>\n\nPlease note that it will take a few days to process your payment.",
   "paymentMethods.manual.instructions.title": "Définir les instructions pour effectuer un transfert vers votre compte bancaire",
   "paymentMethods.manual.upgradePlan": "Abonnez-vous à nos forfaits spéciaux pour les hôtes",
   "paymentMethods.removeConfirm": "Souhaitez-vous vraiment supprimer ce moyen de paiement de votre compte ?",
