+++ lang/pt.json
@@ -1118,7 +1118,6 @@
   "paymentmethod.label": "Método de pagamento",
   "paymentMethod.manual.edit.description": "Os colaboradores poderão escolher \"Transferência bancária\" como forma de pagamento no check-out. As instruções para fazer a transferência eletrônica serão enviadas por e-mail, juntamente com um ID de pedido único. Depois de receber o dinheiro, você poderá marcar o pedido pendente correspondente como pago no painel do organizador.",
   "paymentMethod.manual.edit.description.pricing": "Não há taxa de plataforma para doações feitas dessa maneira. No entanto, solicitamos que você assine nossos planos especiais para administradores fiscais para poder manter e melhorar esse recurso ao longo do tempo (os primeiros US$ 1.000 em orçamento anual estão inclusos no plano gratuito)",
-  "paymentMethod.manual.edit.title": "Permitir que colaboradores façam doações por transferência bancária",
   "paymentMethod.monthlyLimitPerMember.description": "Defina um limite mensal para permitir que outros membros da equipe da sua organização usem esse cartão de crédito dentro desse limite. Se definido como zero, apenas administradores da organização poderão usar este cartão.",
   "paymentMethod.monthlyLimitPerMember.label": "Limite mensal por membro",
   "paymentMethod.success": "Adicionado com sucesso!",
@@ -1139,7 +1138,7 @@
   "paymentMethods.manual.edit": "Altere seus dados bancários",
   "paymentMethods.manual.HowDoesItWork": "Como funciona?",
   "paymentMethods.manual.instructions": "Instruções para transferência bancária",
-  "paymentMethods.manual.instructions.default": "<code>\n     Valor: {amount}\n     Referência/Comunicação: {OrderId}\n     IBAN/Número da conta: BE61310126985517\n     BIC: TRIOBEBB\n     Nome do banco: Triodos\n</code>\n        \nPor favor, note que levará alguns dias para processar seu pagamento.",
+  "paymentMethods.manual.instructions.default": "Please make a bank transfer as follows:\n\n<code>\n    Amount: {amount}\n    Reference/Communication: {OrderId}\n    IBAN/Account Number: BE61310126985517\n    BIC: TRIOBEBB\n    Bank name: Triodos\n</code>\n\nPlease note that it will take a few days to process your payment.",
   "paymentMethods.manual.instructions.title": "Defina as instruções para fazer uma transferência bancária para a sua conta",
   "paymentMethods.manual.upgradePlan": "Inscreva-se nos nossos planos especiais para organizadores",
   "paymentMethods.removeConfirm": "Você deseja realmente remover este método de pagamento da sua conta?",
