+++ lang/ru.json
@@ -1118,7 +1118,6 @@
   "paymentmethod.label": "Метод оплаты",
   "paymentMethod.manual.edit.description": "Contributors will be able to choose \"Bank Transfer\" as a payment method when they check out. The instructions to make the wire transfer will be emailed to them along with a unique order id. Once you received the money, you will be able to mark the corresponding pending order as paid in your host dashboard.",
   "paymentMethod.manual.edit.description.pricing": "There is no platform fee for donations made this way. However, we ask you to kindly subscribe to our special plans for fiscal hosts to be able to maintain and improve this feature over time (the first $1,000 of yearly budget are included in the free plan)",
-  "paymentMethod.manual.edit.title": "Enable contributors to make donations by wire transfer",
   "paymentMethod.monthlyLimitPerMember.description": "Set a monthly limit to allow other team members of your Organization to use this credit card within that limit. If set to zero, only Organization admins will be able to use this card.",
   "paymentMethod.monthlyLimitPerMember.label": "Monthly limit per member",
   "paymentMethod.success": "Successfully added!",
@@ -1139,7 +1138,7 @@
   "paymentMethods.manual.edit": "Edit your bank account details",
   "paymentMethods.manual.HowDoesItWork": "How does it work?",
   "paymentMethods.manual.instructions": "Bank transfer instructions",
-  "paymentMethods.manual.instructions.default": "Please make a bank transfer as follows:\n\n<code>\n    Amount: {amount}\n    Reference/Communication: {OrderId}\n    IBAN/Account Number: BE61310126985517\n    BIC: TRIOBEBB\n    Bank name: Triodos\n</code>\n        \nPlease note that it will take a few days to process your payment.",
+  "paymentMethods.manual.instructions.default": "Please make a bank transfer as follows:\n\n<code>\n    Amount: {amount}\n    Reference/Communication: {OrderId}\n    IBAN/Account Number: BE61310126985517\n    BIC: TRIOBEBB\n    Bank name: Triodos\n</code>\n\nPlease note that it will take a few days to process your payment.",
   "paymentMethods.manual.instructions.title": "Define the instructions to make a bank transfer to your account",
   "paymentMethods.manual.upgradePlan": "Subscribe to our special plans for hosts",
   "paymentMethods.removeConfirm": "Do you really want to remove this payment method from your account?",
